Move Mapping struct to subpackage of replace
